feat: Improve CanFinish function for better accuracy
The code changes modify the CanFinish function in `go/need-for-speed/need_for_speed.go` to improve the accuracy of the calculation. The changes cast the battery, batteryDrain, and speed values to float64 before performing the calculation, ensuring more precise results.

Based on the recent repository commits, the commit message follows the convention of starting with a verb indicating the type of change (feat for a new feature).

+++ go/need-for-speed/need_for_speed.go
@@ -42,5 +42,5 @@ func Drive(car Car) Car {
 
 // CanFinish checks if a car is able to finish a certain track.
 func CanFinish(car Car, track Track) bool {
-	return (car.battery / car.batteryDrain) * car.speed >= track.distance
+	return float64(car.battery / car.batteryDrain) * float64(car.speed) >= float64(track.distance)
 }
